수정 : search, choice 페이지 수정 컨트롤러 추가 spark 다시 시도
