+++ src/main/java/search/service/SearchService.java
@@ -1,5 +1,11 @@
 package search.service;
 
-public class SearchService {
+import java.util.List;
+
+import search.bean.SearchDTO;
+
+public interface SearchService {
+
+    List<SearchDTO> getSearchList(SearchDTO searchDTO);
     
 }
