+++ src/main/webapp/src/components/review/ReviewCSS.css
@@ -6,10 +6,10 @@ body {
 
 .review-container {
     max-width: 800px;
-    margin: 40px auto;
-    padding: 20px;
+    margin: 5px auto;
+    padding-left: 20px;
     background-color: #fff;
-    border: 1px solid #ddd;
+    border: none;
     box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
 }
